+++ game/systems_audiovisual/vertically_flying_number_system.h
@@ -17,10 +17,10 @@ public:
 			vec2 pos;
 		} in;
 
-		float time_of_occurence_seconds = 0.f;
+		double time_of_occurence_seconds = 0.0;
 	};
 
-	float global_time_seconds = 0.f;
+	double global_time_seconds = 0.0;
 
 	std::vector<number> numbers;
